Cast transfer cooldown config value to int
env() returns a string whenever the var is actually set (as
it is once .env.example gets copied to .env in CI), which
Carbon::addDays() rejects with a TypeError. Passed locally
only because the var was unset there, falling through to the
literal int default.

+++ app/Policies/TransferPolicy.php
@@ -26,7 +26,7 @@ class TransferPolicy
             return Response::deny('Your division is not active.');
         }
 
-        $cooldownDays   = config('aod.transfer.cooldown_days');
+        $cooldownDays   = (int) config('aod.transfer.cooldown_days');
         $recentTransfer = Transfer::where('member_id', $user->member_id)
             ->where('created_at', '>=', now()->subDays($cooldownDays))
             ->first();
